// src/auth/providers/interfaces/google-profile.interface.ts
export interface GoogleProfile {
  id: string;
  displayName: string;
  name: {
    familyName: string;
    givenName: string;
  };
  emails: {
    value: string;
    verified: boolean;
  }[];
  photos: {
    value: string;
  }[];
  provider: string;
  _raw: string;
  _json: {
    sub: string;
    name: string;
    given_name: string;
    family_name: string;
    picture: string;
    email: string;
    email_verified: boolean;
  };
}

export interface IGoogleProfile {
  provider: string;
  providerId: string;
  firstName: string;
  lastName: string;
  email: string;
  picture?: string;
}